+++ R/NMdataConf.R
@@ -232,7 +232,8 @@ NMdataConf <- function(...,allow.unknown=FALSE){
             val <- NMdataDecideOption("use.rds",val,allow.unknown=allow.unknown)
             if(!is.null(val)){
                 
-                args1 <- getArgs()
+                ##args1 <- getArgs()
+                args1 <- getArgs(sys.call(),parent.frame())
                 deprecatedArg(oldarg="use.rds",args=args1)
                 message("overwriting `formats.read`")
                 dots[["formats.read"]] <- c("csv")
@@ -461,6 +462,19 @@ NMdataConfOptions <- function(name,allow.unknown=TRUE){
             ## has to be length 1 character or function
            ,is.allowed=function(x) is.function(x) || (length(x)==1 && is.character(x))
            ,msg.not.allowed="file.phi must be a function or a character of length 1"
+           ,process=function(x) {
+               if(is.character(x)) return(function(file) x)
+               x
+           }
+        )
+       ,
+        file.shk=list(
+            default=function(file) {
+                fnExtension(file,ext=".shk")
+            }
+            ## has to be length 1 character or function
+           ,is.allowed=function(x) is.function(x) || (length(x)==1 && is.character(x))
+           ,msg.not.allowed="file.shk must be a function or a character of length 1"
            ,process=function(x) {
                if(is.character(x)) return(function(file) x)
                x