+++ R/NMwriteData.R
@@ -134,7 +134,9 @@ NMwriteData <- function(data,file,formats.write=c("csv","rds"),
         write.RData=FALSE
         write.rds=FALSE
     }
-    args <- getArgs()
+    ## args <- getArgs()
+    
+    args <- getArgs(sys.call(),parent.frame())
     
     args.write.depr <- c("write.rds","write.csv","write.RData")
     if(any(args.write.depr %in% names(args))) {
@@ -192,7 +194,7 @@ NMwriteData <- function(data,file,formats.write=c("csv","rds"),
     ## if args.NMgenText is used, the deprecated ags are not allowed
     ## all.args <- as.list(match.call(expand.dots=FALSE))
     ## These were deprecated way before 2023-06-13
-    all.args <- getArgs()
+    ## all.args <- getArgs()
     if(missing(args.NMgenText)) {
         args.NMgenText <- NULL
     } else {
@@ -201,7 +203,7 @@ NMwriteData <- function(data,file,formats.write=c("csv","rds"),
                        ,fun.msg=stop)
         }
     }
-    used.args.depr <- all.args[names(all.args)%in%args.text.depr]
+    used.args.depr <- args[names(args)%in%args.text.depr]
     if(length(used.args.depr)>0){
         names.used.ad <- names(used.args.depr)
         names.used.ad[names.used.ad=="nm.drop"] <- "drop"
@@ -378,7 +380,7 @@ NMwriteData <- function(data,file,formats.write=c("csv","rds"),
         do.call(write_fst,append(list(x=data,path=file.fst),args.write_fst))
         files.written <- c(files.written,file.fst)
     }
-
+    
     ## write meta data for csv and fst
     if(doStamp && any(c("fst","csv")%in%formats.write)){